fix error when loading transfaces with decimal points

+++ src/lib/regions.py
@@ -432,8 +432,8 @@ class RegionFileInterface:
             region2 = self.indexOfRegionWithName(transData[1])
             faces = []
             for i in range(2, len(transData), 4):
-                p1 = Point(int(transData[i]), int(transData[i+1]))
-                p2 = Point(int(transData[i+2]), int(transData[i+3]))
+                p1 = Point(float(transData[i]), float(transData[i+1]))
+                p2 = Point(float(transData[i+2]), float(transData[i+3]))
                 faces.append(tuple(sorted((p1, p2))))
                 
             # During adjacency matrix reconstruction, we'll mirror over the diagonal
